Renaming, fix missing handlerList, remove useless issuer and make keys accessable
